整理代码;调整目录命名; 实现if语句(if,else,elseif)
